+++ lib/global/event-handlers/onDocumentPointerUp.ts
@@ -4,6 +4,8 @@ import { read, update } from "../mutableState";
 export function onDocumentPointerUp(event: PointerEvent) {
   if (event.defaultPrevented) {
     return;
+  } else if (event.pointerType === "mouse" && event.button > 0) {
+    return;
   }
 
   event.preventDefault();